Device Update

+++ src/main/java/vn/gpay/gsmart/core/api/devices/DevicesTypeRequest.java
@@ -1,5 +1,5 @@
 package vn.gpay.gsmart.core.api.devices;
 
 public class DevicesTypeRequest {
-	public int type;
+	public Long type;
 }
